requirements.txt
undetected-chromedriver
selenium
pandas
matplotlib
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
